# api/xnmo-rosetta-1.4.10-swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "version" : "1.4.10",
    "title" : "Rosetta",
    "description" : "Build Once. Integrate Your Blockchain Everywhere.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
    "/network/list" : {
      "post" : {
        "summary" : "Get List of Available Networks",
        "description" : "This endpoint returns a list of NetworkIdentifiers that the Rosetta server supports.",
        "operationId" : "networkList",
        "tags" : [ "Network" ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MetadataRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Expected response to a valid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NetworkListResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/network/status" : {
      "post" : {
        "summary" : "Get Network Status",
        "description" : "This endpoint returns the current status of the network requested. Any NetworkIdentifier returned by /network/list should be accessible here.",
        "operationId" : "networkStatus",
        "tags" : [ "Network" ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NetworkRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Expected response to a valid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NetworkStatusResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/network/options" : {
      "post" : {
        "summary" : "Get Network Options",
        "description" : "This endpoint returns the version information and allowed network-specific types for a NetworkIdentifier. Any NetworkIdentifier returned by /network/list should be accessible here. Because options are retrievable in the context of a NetworkIdentifier, it is possible to define unique options for each network.",
        "operationId" : "networkOptions",
        "tags" : [ "Network" ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NetworkRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Expected response to a valid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NetworkOptionsResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "NetworkIdentifier" : {
        "description" : "The network_identifier specifies which network a particular object is associated with.",
        "type" : "object",
        "required" : [ "blockchain", "network" ],
        "properties" : {
          "blockchain" : {
            "type" : "string",
            "example" : "bitcoin"
          },
          "network" : {
            "description" : "If a blockchain has a specific chain-id or network identifier, it should go in this field. It is up to the client to determine which network-specific identifier is mainnet or testnet.",
            "type" : "string",
            "example" : "mainnet"
          },
          "sub_network_identifier" : {
            "$ref" : "#/components/schemas/SubNetworkIdentifier"
          }
        }
      },
      "SubNetworkIdentifier" : {
        "description" : "In blockchains with sharded state, the SubNetworkIdentifier is required to query some object on a specific shard. This identifier is optional for all non-sharded blockchains.",
        "type" : "object",
        "required" : [ "network" ],
        "properties" : {
          "network" : {
            "type" : "string",
            "example" : "shard 1"
          },
          "metadata" : {
            "type" : "object",
            "example" : {
              "producer" : "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
            }
          }
        }
      },
      "BlockIdentifier" : {
        "description" : "The block_identifier uniquely identifies a block in a particular network.",
        "type" : "object",
        "required" : [ "index", "hash" ],
        "properties" : {
          "index" : {
            "description" : "This is also known as the block height.",
            "type" : "integer",
            "format" : "int64",
            "example" : 1123941
          },
          "hash" : {
            "type" : "string",
            "example" : "0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85"
          }
        }
      },
...
      "Currency" : {
        "description" : "Currency is composed of a canonical Symbol and Decimals. This Decimals value is used to convert an Amount.Value from atomic units (Satoshis) to standard units (Bitcoins).",
        "type" : "object",
        "required" : [ "symbol", "decimals" ],
        "properties" : {
          "symbol" : {
            "description" : "Canonical symbol associated with a currency.",
            "type" : "string",
            "example" : "BTC"
          },
          "decimals" : {
            "description" : "Number of decimal places in the standard unit representation of the amount. For example, BTC has 8 decimals. Note that it is not possible to represent the value of some currency in atomic units that is not base 10.",
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0,
            "example" : 8
          },
          "metadata" : {
            "description" : "Any additional information related to the currency itself. For example, it would be useful to populate this object with the contract address of an ERC-20 token.",
            "type" : "object",
            "example" : {
              "Issuer" : "Satoshi"
            }
          }
        }
      },
      "SyncStatus" : {
        "description" : "SyncStatus is used to provide additional context about an implementation's sync status. This object is often used by implementations to indicate healthiness when block data cannot be queried until some sync phase completes or cannot be determined by comparing the timestamp of the most recent block with the current time.",
        "type" : "object",
        "properties" : {
          "current_index" : {
            "description" : "CurrentIndex is the index of the last synced block in the current stage. This is a separate field from current_block_identifier in NetworkStatusResponse because blocks with indices up to and including the current_index may not yet be queryable by the caller. To reiterate, all indices up to and including current_block_identifier in NetworkStatusResponse must be queryable via the /block endpoint (excluding indices less than oldest_block_identifier).",
            "type" : "integer",
            "format" : "int64",
            "example" : 100
          },
          "target_index" : {
            "description" : "TargetIndex is the index of the block that the implementation is attempting to sync to in the current stage.",
            "type" : "integer",
            "format" : "int64",
            "example" : 150
          },
          "stage" : {
            "description" : "Stage is the phase of the sync process.",
            "type" : "string",
            "example" : "header sync"
          },
          "synced" : {
            "description" : "sycned is a boolean that indicates if an implementation has synced up to the most recent block. If this field is not populated, the caller should rely on a traditional tip timestamp comparison to determine if an implementation is synced. This field is particularly useful for quiescent blockchains (blocks only produced when there are pending transactions). In these blockchains, the most recent block could have a timestamp far behind the current time but the node could be healthy and at tip.",
            "type" : "boolean"
          }
        }
      },
      "Peer" : {
        "description" : "A Peer is a representation of a node's peer.",
        "type" : "object",
        "required" : [ "peer_id" ],
        "properties" : {
          "peer_id" : {
            "type" : "string",
            "example" : "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
          },
          "metadata" : {
            "type" : "object"
          }
        }
      },
      "Version" : {
        "description" : "The Version object is utilized to inform the client of the versions of different components of the Rosetta implementation.",
        "type" : "object",
        "required" : [ "rosetta_version", "node_version" ],
        "properties" : {
          "rosetta_version" : {
            "description" : "The rosetta_version is the version of the Rosetta interface the implementation adheres to. This can be useful for clients looking to reliably parse responses.",
            "type" : "string",
            "example" : "1.2.5"
          },
          "node_version" : {
            "description" : "The node_version is the canonical version of the node runtime. This can help clients manage deployments.",
            "type" : "string",
            "example" : "1.0.2"
          },
          "middleware_version" : {
            "description" : "When a middleware server is used to adhere to the Rosetta interface, it should return its version here. This can help clients manage deployments.",
            "type" : "string",
            "example" : "0.2.7"
          },
          "metadata" : {
            "description" : "Any other information that may be useful about versioning of dependent services should be returned here.",
            "type" : "object"
          }
        }
      },
      "Allow" : {
        "description" : "Allow specifies supported Operation status, Operation types, and all possible error statuses. This Allow object is used by clients to validate the correctness of a Rosetta Server implementation. It is expected that these clients will error if they receive some response that contains any of the above information that is not specified here.",
        "type" : "object",
        "required" : [ "operation_statuses", "operation_types", "errors", "historical_balance_lookup", "call_methods", "balance_exemptions", "mempool_coins" ],
        "properties" : {
          "operation_statuses" : {
            "description" : "All Operation.Status this implementation supports. Any status that is returned during parsing that is not listed here will cause client validation to error.",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OperationStatus"
            }
          },
          "operation_types" : {
            "description" : "All Operation.Type this implementation supports. Any type that is returned during parsing that is not listed here will cause client validation to error.",
            "type" : "array",
            "items" : {
              "type" : "string",
              "example" : "TRANSFER"
            }
          },
          "errors" : {
            "description" : "All Errors that this implementation could return. Any error that is returned during parsing that is not listed here will cause client validation to error.",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          },
          "historical_balance_lookup" : {
            "type" : "boolean",
            "description" : "Any Rosetta implementation that supports querying the balance of an account at any height in the past should set this to true."
          },
          "timestamp_start_index" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0,
            "description" : "If populated, `timestamp_start_index` indicates the first block index where block timestamps are considered valid (i.e. all blocks less than `timestamp_start_index` could have invalid timestamps). This is useful when the genesis block (or blocks) of a network have timestamp 0. If not populated, block timestamps are assumed to be valid for all available blocks."
          },
          "call_methods" : {
            "type" : "array",
            "description" : "All methods that are supported by the /call endpoint. Communicating which parameters should be provided to /call is the responsibility of the implementer (this is en lieu of defining an entire type system and requiring the implementer to define that in Allow).",
            "items" : {
              "type" : "string",
              "example" : "eth_call"
            }
          },
          "balance_exemptions" : {
            "type" : "array",
            "description" : "BalanceExemptions is an array of BalanceExemption indicating which account balances could change without a corresponding Operation. BalanceExemptions should be used sparingly as they may introduce significant complexity for integrators that attempt to reconcile all account balance changes. If your implementation relies on any BalanceExemptions, you MUST implement historical balance lookup (the ability to query an account balance at any BlockIdentifier).",
            "items" : {
              "$ref" : "#/components/schemas/BalanceExemption"
            }
          },
          "mempool_coins" : {
            "type" : "boolean",
            "description" : "Any Rosetta implementation that can update an AccountIdentifier's unspent coins based on the contents of the mempool should populate this field as true. If false, requests to `/account/coins` that set `include_mempool` as true will be automatically rejected."
          }
        }
      },
      "OperationStatus" : {
        "description" : "OperationStatus is utilized to indicate which Operation status are considered successful.",
        "type" : "object",
        "required" : [ "status", "successful" ],
        "properties" : {
          "status" : {
            "description" : "The status is the network-specific status of the operation.",
            "type" : "string"
          },
          "successful" : {
            "description" : "An Operation is considered successful if the Operation.Amount should affect the Operation.Account. Some blockchains (like Bitcoin) only include successful operations in blocks but other blockchains (like Ethereum) include unsuccessful operations that incur a fee. To reconcile the computed balance from the stream of Operations, it is critical to understand which Operation.Status indicate an Operation is successful and should affect an Account.",
            "type" : "boolean"
          }
        },
        "example" : {
          "status" : "SUCCESS",
          "successful" : true
        }
      },
      "Timestamp" : {
        "description" : "The timestamp of the block in milliseconds since the Unix Epoch. The timestamp is stored in milliseconds because some blockchains produce blocks more often than once a second.",
        "type" : "integer",
        "format" : "int64",
        "minimum" : 0,
        "example" : 1582833600000
      },
...
      "BalanceExemption" : {
        "description" : "BalanceExemption indicates that the balance for an exempt account could change without a corresponding Operation. This typically occurs with staking rewards, vesting balances, and Currencies with a dynamic supply. Currently, it is possible to exempt an account from strict reconciliation by SubAccountIdentifier.Address or by Currency. This means that any account with SubAccountIdentifier.Address would be exempt or any balance of a particular Currency would be exempt, respectively. BalanceExemptions should be used sparingly as they may introduce significant complexity for integrators that attempt to reconcile all account balance changes. If your implementation relies on any BalanceExemptions, you MUST implement historical balance lookup (the ability to query an account balance at any BlockIdentifier).",
        "type" : "object",
        "properties" : {
          "sub_account_address" : {
            "description" : "SubAccountAddress is the SubAccountIdentifier.Address that the BalanceExemption applies to (regardless of the value of SubAccountIdentifier.Metadata).",
            "type" : "string",
            "example" : "staking"
          },
          "currency" : {
            "$ref" : "#/components/schemas/Currency"
          },
          "exemption_type" : {
            "$ref" : "#/components/schemas/ExemptionType"
          }
        }
      },
      "ExemptionType" : {
        "description" : "ExemptionType is used to indicate if the live balance for an account subject to a BalanceExemption could increase above, decrease below, or equal the computed balance. * greater_or_equal: The live balance may increase above or equal the computed balance. This typically   occurs with staking rewards that accrue on each block. * less_or_equal: The live balance may decrease below or equal the computed balance. This typically   occurs as balance moves from locked to spendable on a vesting account. * dynamic: The live balance may increase above, decrease below, or equal the computed balance. This   typically occurs with tokens that have a dynamic supply.",
        "type" : "string",
        "enum" : [ "greater_or_equal", "less_or_equal", "dynamic" ]
      },
...
      "MetadataRequest" : {
        "description" : "A MetadataRequest is utilized in any request where the only argument is optional metadata.",
        "type" : "object",
        "properties" : {
          "metadata" : {
            "type" : "object"
          }
        }
      },
      "NetworkListResponse" : {
        "description" : "A NetworkListResponse contains all NetworkIdentifiers that the node can serve information for.",
        "type" : "object",
        "required" : [ "network_identifiers" ],
        "properties" : {
          "network_identifiers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NetworkIdentifier"
            }
          }
        }
      },
      "NetworkRequest" : {
        "description" : "A NetworkRequest is utilized to retrieve some data specific exclusively to a NetworkIdentifier.",
        "type" : "object",
        "required" : [ "network_identifier" ],
        "properties" : {
          "network_identifier" : {
            "$ref" : "#/components/schemas/NetworkIdentifier"
          },
          "metadata" : {
            "type" : "object"
          }
        }
      },
      "NetworkStatusResponse" : {
        "description" : "NetworkStatusResponse contains basic information about the node's view of a blockchain network. It is assumed that any BlockIdentifier.Index less than or equal to CurrentBlockIdentifier.Index can be queried. If a Rosetta implementation prunes historical state, it should populate the optional `oldest_block_identifier` field with the oldest block available to query. If this is not populated, it is assumed that the `genesis_block_identifier` is the oldest queryable block. If a Rosetta implementation performs some pre-sync before it is possible to query blocks, sync_status should be populated so that clients can still monitor healthiness. Without this field, it may appear that the implementation is stuck syncing and needs to be terminated.",
        "type" : "object",
        "required" : [ "current_block_identifier", "current_block_timestamp", "genesis_block_identifier", "peers" ],
        "properties" : {
          "current_block_identifier" : {
            "$ref" : "#/components/schemas/BlockIdentifier"
          },
          "current_block_timestamp" : {
            "$ref" : "#/components/schemas/Timestamp"
          },
          "genesis_block_identifier" : {
            "$ref" : "#/components/schemas/BlockIdentifier"
          },
          "oldest_block_identifier" : {
            "$ref" : "#/components/schemas/BlockIdentifier"
          },
          "sync_status" : {
            "$ref" : "#/components/schemas/SyncStatus"
          },
          "peers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Peer"
            }
          }
        }
      },
      "NetworkOptionsResponse" : {
        "description" : "NetworkOptionsResponse contains information about the versioning of the node and the allowed operation statuses, operation types, and errors.",
        "type" : "object",
        "required" : [ "version", "allow" ],
        "properties" : {
          "version" : {
            "$ref" : "#/components/schemas/Version"
          },
          "allow" : {
            "$ref" : "#/components/schemas/Allow"
          }
        }
      },
...
      "Error" : {
        "description" : "Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.",
        "type" : "object",
        "required" : [ "code", "message", "retriable" ],
        "properties" : {
          "code" : {
            "description" : "Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.",
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0,
            "example" : 12
          },
          "message" : {
            "description" : "Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.",
            "type" : "string",
            "example" : "Invalid account format"
          },
          "description" : {
            "description" : "Description allows the implementer to optionally provide additional information about an error. In many cases, the content of this field will be a copy-and-paste from existing developer documentation. Description can ONLY be populated with generic information about a particular type of error. It MUST NOT be populated with information about a particular instantiation of an error (use `details` for this). Whereas the content of Error.Message should stay stable across releases, the content of Error.Description will likely change across releases (as implementers improve error documentation). For this reason, the content in this field is not part of any type assertion (unlike Error.Message).",
            "type" : "string",
            "example" : "This error is returned when the requested AccountIdentifier is improperly formatted."
          },
          "retriable" : {
            "description" : "An error is retriable if the same request may succeed if submitted again.",
            "type" : "boolean"
          },
          "details" : {
            "type" : "object",
            "description" : "Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.",
            "example" : {
              "address" : "0x1dcc4de8dec75d7aab85b567b6",
              "error" : "not base64"
            }
          }
        }
      }
    }
  }
}