# server/go-server-server-generated/go/model_network_list_response.go
/*
 * Rosetta
 *
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * API version: 1.4.10
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// A NetworkListResponse contains all NetworkIdentifiers that the node can serve information for.
type NetworkListResponse struct {

	NetworkIdentifiers []NetworkIdentifier `json:"network_identifiers"`
}